pyproject.toml
[project]
name = "ifb-profi"
version = "0.1.0"
description = "Automatisierte Antragsprüfung für IFB PROFI - Option 1 (Super-Lite MVP)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Patrick Zackert", email = "patrick.zackert@example.com" }
]
license = { text = "MIT" }

dependencies = [
    # UI
    "streamlit>=1.28.0",
    
    # Document Parsing
    "pymupdf>=1.23.0",      # PDF (PyMuPDF)
    "python-docx>=1.0.0",   # DOCX
    "openpyxl>=3.1.0",      # XLSX
    
    # RAG System
    "chromadb>=0.4.15",                                           # Vector Store
    "sentence-transformers>=2.2.0",                               # Embeddings
    
    # LLM Client
    "openai>=1.3.0",        # OpenAI-kompatible API für LM Studio
    
    # Utils
    "pyyaml>=6.0.0",        # Config
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
